1 array iterator

+++ function_pointers/1-array_iterator.c
@@ -2,6 +2,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
+
 /**
 * array_iterator - executes a function given as a parameter on each element of an array
 * @array: array
@@ -10,7 +11,7 @@
 */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-int i;
+size_t i;
 for (i = 0; i < size; i++)
 {
 action(array[i]);
